Move face functions into Character
