+++ game/src/main/kotlin/world/gregs/voidps/world/activity/skill/runecrafting/EssenceMine.kt
@@ -2,7 +2,6 @@ package world.gregs.voidps.world.activity.skill.runecrafting
 
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.data.definition.AreaDefinitions
-import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.character.move.tele
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.entity.character.player.Player
