+++ game/src/main/kotlin/world/gregs/voidps/world/activity/skill/smithing/Cannonballs.kts
@@ -3,7 +3,6 @@ package world.gregs.voidps.world.activity.skill.smithing
 import com.github.michaelbull.logging.InlineLogger
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.client.ui.interact.itemOnObjectOperate
-import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.character.mode.interact.Interaction
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.chat.ChatType
