+++ game/src/main/kotlin/world/gregs/voidps/world/command/debug/PlayerUpdatingCommands.kts
@@ -10,6 +10,7 @@ import world.gregs.voidps.engine.entity.character.player.*
 import world.gregs.voidps.engine.get
 import world.gregs.voidps.engine.inject
 import world.gregs.voidps.engine.map.zone.DynamicZones
+import world.gregs.voidps.type.Delta
 import world.gregs.voidps.type.Direction
 import world.gregs.voidps.world.interact.entity.combat.hit.damage
 import world.gregs.voidps.world.interact.entity.effect.transform
@@ -104,7 +105,7 @@ adminCommand("shoot") {
 adminCommand("face (delta-x) (delta-y)", "turn player to face a direction or delta coordinate") {
     if (content.contains(" ")) {
         val parts = content.split(" ")
-        player.face(parts[0].toInt(), parts[1].toInt())
+        player.face(Delta(parts[0].toInt(), parts[1].toInt()))
     } else {
         val direction = Direction.valueOf(content.toScreamingSnakeCase())
         player.face(direction.delta)
