+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/item/ItemPickup.kts
@@ -3,7 +3,6 @@ package world.gregs.voidps.world.interact.entity.item
 import com.github.michaelbull.logging.InlineLogger
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.entity.character.player.chat.inventoryFull
-import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.item.floor.FloorItems
 import world.gregs.voidps.engine.entity.item.floor.floorItemOperate
 import world.gregs.voidps.engine.entity.item.floor.npcOperateFloorItem
