+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/Login.kts
@@ -4,7 +4,6 @@ import world.gregs.voidps.bot.isBot
 import world.gregs.voidps.engine.entity.character.player.flagAppearance
 import world.gregs.voidps.engine.entity.character.player.flagMovementType
 import world.gregs.voidps.engine.entity.character.player.flagTemporaryMoveType
-import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.playerSpawn
 
 playerSpawn { player ->
@@ -15,6 +14,6 @@ playerSpawn { player ->
         player.flagMovementType()
         player.flagTemporaryMoveType()
         player.flagAppearance()
-        player.face()
+        player.clearFace()
     }
 }
