+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/Attack.kts
@@ -2,7 +2,6 @@ package world.gregs.voidps.world.interact.entity.player.combat
 
 import world.gregs.voidps.engine.client.ui.interact.itemOnNPCApproach
 import world.gregs.voidps.engine.entity.character.Character
-import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.character.mode.EmptyMode
 import world.gregs.voidps.engine.entity.character.mode.interact.Interact
 import world.gregs.voidps.engine.entity.character.npc.characterApproachNPC
