+++ game/src/main/kotlin/world/gregs/voidps/world/map/ourania/ZamorakCrafter.kts
@@ -1,7 +1,6 @@
 package world.gregs.voidps.world.map.ourania
 
 import world.gregs.voidps.engine.data.definition.PatrolDefinitions
-import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.character.mode.Patrol
 import world.gregs.voidps.engine.entity.character.mode.move.npcMove
 import world.gregs.voidps.engine.entity.npcSpawn
