+++ game/src/main/kotlin/world/gregs/voidps/world/map/varrock/palace/SirPrysin.kts
@@ -1,7 +1,6 @@
 package world.gregs.voidps.world.map.varrock.palace
 
 import world.gregs.voidps.engine.entity.character.clearWatch
-import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.character.mode.PauseMode
 import world.gregs.voidps.engine.entity.character.move.tele
 import world.gregs.voidps.engine.entity.character.npc.NPCOption
